JSIP-416 - small patch

+++ branches/JAIN-SIP-NIO-year-2012/src/gov/nist/javax/sip/stack/SIPTransactionStack.java
@@ -406,7 +406,7 @@ public abstract class SIPTransactionStack implements
     // factory used to create MessageProcessor objects
     public MessageProcessorFactory messageProcessorFactory;
     
-    public int nioSocketMaxIdleTime;
+    public long nioSocketMaxIdleTime;
 
     protected boolean aggressiveCleanup = false;
